==> 3_foisweep/3_plots.Rmd <==
---
title: "3_plots"
output: html_document
date: "2023-03-15"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```



##Compiling time series of cases, seroprev and immunity for facet illustration
##Cases per capita
```{r}
inc0<-readRDS("0_res/novax_inc.RDS")
inc1 <- readRDS("0_res/vaxsero_inc.RDS")
inc2 <- readRDS("0_res/vaxint_inc.RDS")
inc<- rbind(inc0,inc1, inc2%>%filter(kappa==2500))

inc<-inc%>%filter(sero_thresh %in% c("0","0.5","0.6","0.7","0.8","ann_435", "bi_435"))%>%
      mutate(inc_all = (inc_all/tot_pop)*100)

med<- inc%>%group_by(date,sero_thresh,foi_sp)%>%
      dplyr::summarise(med = median(inc_all),
                prob_25 = quantile(inc_all, probs=0.10),
                prob_75 = quantile(inc_all, probs=0.90))

p1 <- med %>% ggplot()+
    geom_ribbon(aes(x=date, ymin=prob_25, ymax=prob_75, fill="Range"))+
     theme_bw()+geom_line(aes(x=date, y=med, colour = "Cases \n per 100"), size=0.6)+  

    scale_colour_manual("",values="#EF3B2C")+
    scale_fill_manual("",values="gray78")+
  #ggtitle("10-year epidemic projection")+
  xlab("") + ylab("Cases per 100")+
  facet_grid(vars(sero_thresh),vars(foi_sp))+
  #facet_wrap(~sero_thresh,nrow=1)+
  plot_theme


#png("0_plot/9_inc_strip.png",width = 12,height = 12,units="in",res=400)
p1
#dev.off()


```

## Deaths per capita
```{r}
death0 <- readRDS("0_res/novax_death.RDS")
death1 <- readRDS("0_res/vaxsero_death.RDS")
death2 <- readRDS("0_res/vaxint_death.RDS")
death<- rbind(death0,death1, death2)%>%
        filter(sero_thresh %in% c("0","0.5","0.6","0.7","0.8","ann_late", "bi_late"))%>%
        filter(kappa==2500)

med<- death%>%
     mutate(deathpermil = (new_Deaths_tot/tot_pop)*1000000)%>%
      group_by(date,sero_thresh,foi_sp)%>%
      dplyr::summarise(med = median(deathpermil),
                prob_25 = quantile(deathpermil, probs=0.10),
                prob_75 = quantile(deathpermil, probs=0.90))

p9 <- med %>% ggplot()+
    geom_ribbon(aes(x=date, ymin=prob_25, ymax=prob_75, fill="Range"))+
     theme_bw()+geom_line(aes(x=date, y=med, colour = "Cases \n per 100"), size=0.6)+  

    scale_colour_manual("",values="#EF3B2C")+
    scale_fill_manual("",values="gray78")+
  #ggtitle("10-year epidemic projection")+
  xlab("") + ylab("Deaths per million")+
  #facet_wrap(~sero_thresh, nrow=1)+
  facet_grid(vars(sero_thresh),vars(foi_sp))+

  plot_theme




#png("0_plot/9_death_strip.png",width = 12,height = 5,units="in",res=300)
p9
#dev.off()

```



## Sero facet wrap
```{r}
sero0 <- readRDS("0_res/novax_sero.RDS")
sero1 <- readRDS("0_res/vaxsero_sero.RDS")
sero2 <- readRDS("0_res/vaxint_sero.RDS")
sero<- rbind(sero0,sero1, sero2)


thresh_line <- data.frame(scen=c("No vax","50% thresh", "65% thresh","80% thresh","Annual","Biennial"),
                          sero_thresh = c("0","0.5","0.65","0.8","ann_late","bi_late"),
                          thresh_num = c(NA,51,66,80,NA,NA))%>%
              mutate(scen = factor(scen, levels = c("No vax", "50% thresh", "65% thresh", "80% thresh", "Annual","Biennial")))

sero<-sero%>%filter( sero_thresh %in% c("0","0.5","0.65","0.8","ann_late", "bi_late"))
sero<-sero%>%select(-sero_all)%>%pivot_longer(cols = sero_c:sero_e, names_to = "sero",values_to="val")

sero_med<- sero%>%group_by(date,sero_thresh,sero,foi_sp)%>%
      dplyr::summarise(med = median(val*100),
                prob25 = quantile(val*100, probs=0.25),
                prob75 = quantile(val*100, probs=0.75)) %>%
      left_join(thresh_line %>%select(scen, sero_thresh))

sero_med <- sero_med%>%mutate(
              scen = factor(scen, levels = c("No vax", "50% thresh", "65% thresh", "80% thresh", "Annual","Biennial"))
              )
sero_range_c <- sero_med %>%filter(sero=="sero_c")
sero_range_a <- sero_med %>%filter(sero=="sero_a")
sero_range_e <- sero_med %>%filter(sero=="sero_e")
                

p2 <-  ggplot()+
      geom_ribbon(data=sero_range_c,aes(x=date, ymin=prob25, ymax=prob75), fill="gray78",alpha=0.5)+
      geom_ribbon(data=sero_range_a,aes(x=date, ymin=prob25, ymax=prob75), fill="gray78", alpha=0.5)+
      geom_ribbon(data=sero_range_e,aes(x=date, ymin=prob25, ymax=prob75), fill="gray78", alpha=0.5)+
      geom_line(data=sero_med,aes(x=date, y = med, col = sero),size=1)+       
       geom_hline(data=thresh_line,aes(yintercept=thresh_num),linetype="dashed",size=0.6)+
  facet_grid(vars(sero_thresh),vars(foi_sp))+
      ylim(0,100) +theme_bw()+
      scale_color_manual(values = c("#4292C6","#C6DBEF","#08306B"),
                                   labels=c("Adult",
                                            "Child",
                                            "Older \nadults"))+
        xlab("")+
                theme(axis.ticks.x = element_blank(), axis.text.x = element_blank(),
                      axis.text.y = element_text(size=11),
                      strip.text = element_text(size=13),
                      legend.position = "right",legend.title = element_blank(), legend.text = element_text(size=12),
                      plot.margin = unit(c(0,0,-0.5,1), "line"))+
 
  ylab("Seroprevalence\n by age(%)")+
  theme(panel.background = element_rect(fill = "white", colour = "black",size = 0.7, linetype = "solid"),
        panel.grid.major = element_line(size = 0.00001, linetype = 'solid',colour = "white"),
        panel.grid.minor = element_line(size = 0.00001, linetype = 'solid',
                                colour = "white"))

#png("0_plot/9_sero_strip.png",width = 12,height = 12,units="in",res=400)
p2
#dev.off()

sero_med <- sero_med%>%mutate(
              scen = factor(scen, levels = c("No vax", "50% thresh", "65% thresh", "80% thresh", "Annual","Biennial")))%>%filter(foi_sp==0.7)
sero_range_c <- sero_med %>%filter(sero=="sero_c")
sero_range_a <- sero_med %>%filter(sero=="sero_a")
sero_range_e <- sero_med %>%filter(sero=="sero_e")
                

 ggplot()+
      geom_ribbon(data=sero_range_c,aes(x=date, ymin=prob25, ymax=prob75), fill="gray78",alpha=0.5)+
      geom_ribbon(data=sero_range_a,aes(x=date, ymin=prob25, ymax=prob75), fill="gray78", alpha=0.5)+
      geom_ribbon(data=sero_range_e,aes(x=date, ymin=prob25, ymax=prob75), fill="gray78", alpha=0.5)+
      geom_line(data=sero_med,aes(x=date, y = med, col = sero),size=1)+       
       geom_hline(data=thresh_line,aes(yintercept=thresh_num),linetype="dashed",size=0.6)+
      facet_wrap(~sero_thresh, nrow=1)+
      ylim(0,100) +theme_bw()+
      scale_color_manual(values = c("#4292C6","#C6DBEF","#08306B"),
                                   labels=c("Adult",
                                            "Child",
                                            "Older \nadults"))+
        xlab("")+
                theme(axis.ticks.x = element_blank(), axis.text.x = element_blank(),
                      axis.text.y = element_text(size=11),
                      strip.text = element_text(size=13),
                      legend.position = "right",legend.title = element_blank(), legend.text = element_text(size=12),
                      plot.margin = unit(c(0,0,-0.5,1), "line"))+
 
  ylab("Seroprevalence\n by age(%)")+
  theme(panel.background = element_rect(fill = "white", colour = "black",size = 0.7, linetype = "solid"),
        panel.grid.major = element_line(size = 0.00001, linetype = 'solid',colour = "white"),
        panel.grid.minor = element_line(size = 0.00001, linetype = 'solid',
                                colour = "white"))
```

```{r}
imm0 <- readRDS("0_res/novax_imm_med.RDS")
imm1 <- readRDS("0_res/vaxsero_imm_med.RDS")
imm2 <- readRDS("0_res/vaxint_imm_med.RDS")

imm_med<- rbind(imm0,imm1,imm2)%>%
            filter(sero_thresh%in%c("0","0.5","0.6","0.7","0.8","ann_late", "bi_late"))

imm_med <- imm_med %>%
            left_join(data.frame(age_grp = c("a","c","e"),
                                  age_lab = c("Adult","Child","Older adult")))%>%
            left_join(data.frame(sero_thresh = c("0","0.5","0.6","0.7","0.8","ann_late","bi_late"),
                                  sero_lab = c("No vax","50% thresh", "60% thresh", "70% thresh","80% thresh","Annual","Biennial")))%>%
            #left_join(data.frame(sero_thresh = c("0","0.5","0.65","0.8","ann_late","bi_late"),
            #                     sero_lab = c("No vax","50% thresh", "65% thresh", "80% thresh","Annual","Biennial")))%>%
            mutate(age_lab = factor(age_lab, levels=c("Child","Adult","Older adult")),
                   sero_lab = factor(sero_lab, levels =  c("No vax","50% thresh", "60% thresh", "70% thresh","80% thresh","Annual","Biennial")))





##Immune landscape for older adults only
imm_med_kappa2500 <- imm_med%>%filter(kappa==2500)

p4 <- imm_med_kappa2500%>%
          filter(age_grp == "e") %>%
          ggplot(aes(x=date, y=med, fill=imm_cat))+
          geom_col(aes(x=date, y=med, fill=imm_cat, width=1), position="fill")+
          scale_fill_manual(values = c("#FFFFCC",
                                       "#35978F","#80CDC1","#C7EAE5",
                                       brewer.pal(9, "PuBu")[c(7,6,5,4,3,2)]))+
            facet_grid(vars(sero_thresh),vars(foi_sp))+
       theme_classic()+
          xlab("Year") + ylab("Population proportion")+
    theme(plot.title = element_text(size=14),
                      axis.text.x = element_text(angle=30, vjust=1.2, hjust=1, size=11),
                      axis.text = element_text(size=11),
                      axis.title = element_text(size=12),
                      #strip.text = element_blank(),
                      legend.text = element_text(size=12),
                      legend.position = "right",
                      plot.margin = unit(c(-0.5,0.2,-1,1), 'lines'))+
  theme(panel.background = element_rect(fill = "white", colour = "black",size = 0.7, linetype = "solid"),
        panel.grid.major = element_line(size = 0.001, linetype = 'solid',colour = "white"),
        panel.grid.minor = element_line(size = 0.001, linetype = 'solid',
                                colour = "white"))



#png("0_plot/imm_strip_simp.png",width = 12,height = 14,units="in",res=400)
#p3
#dev.off()

#png("0_plot/imm_stripold.png",width = 13,height = 3,units="in",res=400)
p4
#dev.off()

```
```{r}
##Explore imm_med...
try<- imm_med_kappa2500%>%
  filter(foi_sp==0.8&age_grp=="e")%>%
  filter(date%in% c(as.Date("2023-05-01"),
                    as.Date("2025-08-01"),
                    as.Date("2027-10-01"),
                    as.Date("2029-12-01"),
                    as.Date("2031-07-01")))
##Singular run to check..
imm_9<- readRDS("0_res/vaxsero_imm.RDS")

imm_9<- imm_9%>%filter(sweep_unique.y==10)%>%
            filter(sero_thresh%in%c("0","0.5","0.6","0.7","0.8","ann_late", "bi_late"))

imm9<- imm_9 %>%
            left_join(data.frame(age_grp = c("a","c","e"),
                                  age_lab = c("Adult","Child","Older adult")))%>%
            left_join(data.frame(sero_thresh = c("0","0.5","0.6","0.7","0.8","ann_late","bi_late"),
                                  sero_lab = c("No vax","50% thresh", "60% thresh", "70% thresh","80% thresh","Annual","Biennial")))%>%
            #left_join(data.frame(sero_thresh = c("0","0.5","0.65","0.8","ann_late","bi_late"),
            #                     sero_lab = c("No vax","50% thresh", "65% thresh", "80% thresh","Annual","Biennial")))%>%
            mutate(age_lab = factor(age_lab, levels=c("Child","Adult","Older adult")),
                   sero_lab = factor(sero_lab, levels =  c("No vax","50% thresh", "60% thresh", "70% thresh","80% thresh","Annual","Biennial")))



imm9%>% filter(age_grp == "e") %>%
          ggplot(aes(x=date, y=val, fill=imm_cat))+
          geom_col(aes(x=date, y=val, fill=imm_cat, width=1), position="fill")+
          scale_fill_manual(values = c("#FFFFCC",
                                       "#35978F","#80CDC1","#C7EAE5",
                                       brewer.pal(9, "PuBu")[c(7,6,5,4,3,2)]))+
            facet_grid(vars(sero_thresh),vars(foi_sp))+
       theme_classic()+
          xlab("Year") + ylab("Population proportion")+
    theme(plot.title = element_text(size=14),
                      axis.text.x = element_text(angle=30, vjust=1.2, hjust=1, size=11),
                      axis.text = element_text(size=11),
                      axis.title = element_text(size=12),
                      #strip.text = element_blank(),
                      legend.text = element_text(size=12),
                      legend.position = "right",
                      plot.margin = unit(c(-0.5,0.2,-1,1), 'lines'))+
  theme(panel.background = element_rect(fill = "white", colour = "black",size = 0.7, linetype = "solid"),
        panel.grid.major = element_line(size = 0.001, linetype = 'solid',colour = "white"),
        panel.grid.minor = element_line(size = 0.001, linetype = 'solid',
                                colour = "white"))

```



```{r}
p3 <-ggarrange(p2,p1,p4,ncol=1, align="v")



png("0_plot/2_fig2_hiwane.png", width = 11,height = 11,units="in",res=600)
as_ggplot(arrangeGrob(p3, p11o,p12o, nrow=7,ncol =2, 
                      layout_matrix=rbind(c(1,1), c(1,1),c(1,1),c(1,1),c(2,3),c(2,3),c(2,3))))
dev.off()


```
